+++ lending/CodatLending/ManageDataPullOperations.cs
@@ -35,15 +35,7 @@ namespace CodatLending
         /// List pull operations
         /// 
         /// <remarks>
-        /// The *List pull operations* endpoint returns a list of <a href="https://docs.codat.io/lending-api#/schemas/PullOperation">pull operations</a> made by your client.<br/>
-        /// <br/>
-        /// A <a href="https://docs.codat.io/lending-api#/schemas/PullOperation">pull operation</a> is a request to retrieve a specific data type from an integration.<br/>
-        /// <br/>
-        /// ### Tips and traps<br/>
-        /// <br/>
-        /// - The *List pull operations* endpoint does not support querying the `isCompleted` property. You can filter failed pull operations by querying `status!=Complete&amp;&amp;status!=NotSupported` instead.<br/>
-        /// <br/>
-        /// 
+        /// Gets the pull operation history (datasets) for a given company.
         /// </remarks>
         /// </summary>
         Task<ListPullOperationsResponse> ListAsync(ListPullOperationsRequest? request = null);
@@ -53,10 +45,10 @@ namespace CodatLending
     {
         public SDKConfig Config { get; private set; }
         private const string _language = "csharp";
-        private const string _sdkVersion = "4.4.0";
-        private const string _sdkGenVersion = "2.169.0";
+        private const string _sdkVersion = "4.4.1";
+        private const string _sdkGenVersion = "2.173.0";
         private const string _openapiDocVersion = "3.0.0";
-        private const string _userAgent = "speakeasy-sdk/csharp 4.4.0 2.169.0 3.0.0 Codat.Lending";
+        private const string _userAgent = "speakeasy-sdk/csharp 4.4.1 2.173.0 3.0.0 Codat.Lending";
         private string _serverUrl = "";
         private ISpeakeasyHttpClient _defaultClient;
         private ISpeakeasyHttpClient _securityClient;
@@ -106,7 +98,7 @@ namespace CodatLending
                 
                 return response;
             }
-            if((response.StatusCode == 401) || (response.StatusCode == 404) || (response.StatusCode == 429))
+            if((response.StatusCode == 401) || (response.StatusCode == 402) || (response.StatusCode == 403) || (response.StatusCode == 404) || (response.StatusCode == 429) || (response.StatusCode == 500) || (response.StatusCode == 503))
             {
                 if(Utilities.IsContentTypeMatch("application/json",response.ContentType))
                 {
@@ -155,7 +147,7 @@ namespace CodatLending
                 
                 return response;
             }
-            if((response.StatusCode == 400) || (response.StatusCode == 401) || (response.StatusCode == 404) || (response.StatusCode == 429))
+            if((response.StatusCode == 400) || (response.StatusCode == 401) || (response.StatusCode == 402) || (response.StatusCode == 403) || (response.StatusCode == 404) || (response.StatusCode == 429) || (response.StatusCode == 500) || (response.StatusCode == 503))
             {
                 if(Utilities.IsContentTypeMatch("application/json",response.ContentType))
                 {