+++ lending/CodatLending/Models/Operations/GetCommerceLifetimeValueMetricsResponse.cs
@@ -28,7 +28,7 @@ namespace CodatLending.Models.Operations
         public string? ContentType { get; set; } = default!;
 
         /// <summary>
-        /// Your API request was not properly authorized.
+        /// The request made is not valid.
         /// </summary>
         public ErrorMessage? ErrorMessage { get; set; }