+++ lending/RELEASES.md
@@ -163,3 +163,13 @@ Based on:
 - [csharp v4.4.0] lending
 ### Releases
 - [NuGet v4.4.0] https://www.nuget.org/packages/Codat.Lending/4.4.0 - lending
+
+## 2023-10-30 01:26:08
+### Changes
+Based on:
+- OpenAPI Doc 3.0.0 https://raw.githubusercontent.com/codatio/oas/main/yaml/Codat-Lending.yaml
+- Speakeasy CLI 1.109.0 (2.173.0) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [csharp v4.4.1] lending
+### Releases
+- [NuGet v4.4.1] https://www.nuget.org/packages/Codat.Lending/4.4.1 - lending